% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Between_variance.R
\name{Between_variance}
\alias{Between_variance}
\title{Significance of the between time-point variance}
\usage{
Between_variance(data, Features, Time_Points)
}
\arguments{
\item{data}{Longitudinal data}

\item{Features}{Vector containing the names of the features}

\item{Time_Points}{Vector containing the time-point of each cell}
}
\value{
A list containing the p-values of each feature between different time-points
}
\description{
This function calculates the significance of the variance between time-points
}
