% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_state_order.R
\name{cell_state_order}
\alias{cell_state_order}
\title{Ordering the cell states}
\usage{
cell_state_order(data, root, cluster, feat_num, dist_method = "minkowski")
}
\arguments{
\item{data}{Your data (one time point)}

\item{root}{The starting cell state of the ordering process}

\item{cluster}{Name of the cluster column}

\item{feat_num}{Number of features}

\item{dist_method}{The distance method used to generate the distance matrix}
}
\value{
A vector with the ordered clusters
}
\description{
This function orders the cell states inside one time point of the biological process based on the 
distances between cell states
}
