% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_expr_net.R
\name{co_expr_net}
\alias{co_expr_net}
\title{Plot co-expression network}
\usage{
co_expr_net(data, cut_corr = 0.3, main)
}
\arguments{
\item{data}{The feature expression data}

\item{cut_corr}{Plot features with correlation higher than}

\item{main}{The title of the plot}
}
\value{
A co-expression network
}
\description{
This function plots the co-expression network of the features
}
