% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_tp.R
\name{compare_tp}
\alias{compare_tp}
\title{Feature expression between time-points}
\usage{
compare_tp(Data, Cell_Type, Time_Step, Feature = "", main = "")
}
\arguments{
\item{Data}{Data.frame of feature expressions}

\item{Cell_Type}{Cell-type of each input cell}

\item{Time_Step}{Time-point of each input cell.}

\item{Feature}{The feature which expression will be plotted}

\item{main}{Title of plot}
}
\value{
A plot comparing the expression of a feature at different time-points.
}
\description{
This function generates a plot comparing the expression of a feature at different time-points.
}
