% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation_lstm.R
\name{CrossValidation_lstm}
\alias{CrossValidation_lstm}
\title{Cross-Validation of lstm model}
\usage{
CrossValidation_lstm(
  data,
  Feature,
  periods_train = 300,
  periods_test = 100,
  skip_span = 10,
  batch_size = 50,
  tsteps = 1,
  epochs = 1000
)
}
\arguments{
\item{data}{Your time step data}

\item{Feature}{Interested feature}

\item{periods_train}{Number of training time steps}

\item{periods_test}{Number of testing time steps}

\item{skip_span}{Distribute samples into n sets that span the entire historical data}

\item{batch_size}{The number of training examples in one forward/backward pass of a 
RNN before a weight update. It must be evenly divisible into both the training and testing lengths}

\item{tsteps}{The number of lags included in the training/testing set}

\item{epochs}{The total number of forward/backward pass iterations}
}
\value{
A dataframe with both original and predicted testing time steps
}
\description{
This function runs a cross-validation lstm model to check if your parameters are 
appropriate for the forecasting of future trends of the feature
}
