% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divHier.R
\name{divHier}
\alias{divHier}
\title{Divisive hierarchical clustering}
\usage{
divHier(data, k = 200, dist_method = "minkowski", clust_method = "ward.D2")
}
\arguments{
\item{data}{Your data (must be splitted by time points)}

\item{k}{Number of clusters (default=200)}

\item{dist_method}{The method to generate the distance matrix for clustering (default="minkowski")}

\item{clust_method}{The method to clusterize the data (default="ward.D2")}
}
\value{
The original data with a column of clusters
}
\description{
This function applies a hierarchical clustering to each time point of your data,
generating cluster considered as cell states.
}
