% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expr_Heatmap.R
\name{Expr_Heatmap}
\alias{Expr_Heatmap}
\title{Expression heatmap}
\usage{
Expr_Heatmap(data, Features)
}
\arguments{
\item{data}{Entire data.frame}

\item{Features}{The selected features}
}
\value{
A list containing the heatmaps
}
\description{
This function plots several heatmaps representing the expression of the genes in
different cells at different time-points, taking into account the cell-type.
}
