% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactFeat_corr.R
\name{FactFeat_corr}
\alias{FactFeat_corr}
\title{Factor - Feature correlation}
\usage{
FactFeat_corr(Feat_data, Fact_data, Features, positive = F)
}
\arguments{
\item{Feat_data}{Data.frame of feature expressions}

\item{Fact_data}{Data.frame of factor expression}

\item{Features}{Vector of feature names}

\item{positive}{Obtain the positive correlated features? (Default=FALSE)}
}
\value{
A list containing the Factor-Feature correlation or a vector
containing the positive correlated features.
}
\description{
This function calculates the correlation between each feature and the factor.
You can obtain the positive correlated features (with the factor) by setting
"positive" to true.
}
