% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactFeat_Dyn.R
\name{FactFeat_Dyn}
\alias{FactFeat_Dyn}
\title{Factor-Feature dynamics}
\usage{
FactFeat_Dyn(data, Features, Fact_data, tp_explicit = TRUE)
}
\arguments{
\item{data}{Entire data.frame}

\item{Features}{The selected features and the factor}

\item{Fact_data}{Factor expression data}

\item{tp_explicit}{Take into account time_points?}
}
\value{
A plot representing the dynamic change of feature and
factor expressions.
}
\description{
This function generates a plot representing the dynamic change of feature and
factor expressions.
}
