% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactFeat_ExprRegr.R
\name{FactFeat_ExprRegr}
\alias{FactFeat_ExprRegr}
\title{Factor-Feature expression regression}
\usage{
FactFeat_ExprRegr(Feat_data, Fact_data, Features, y = "Factor")
}
\arguments{
\item{Feat_data}{Data.frame of feature expressions}

\item{Fact_data}{Data.frame of factor expression}

\item{Features}{Vector of feature names}

\item{y}{Name y-axis}

\item{x}{Name x-axis}
}
\value{
A list of plots
}
\description{
This function generates a list of plots comparing each feature with the same factor
}
