% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feat_plot.R
\name{feat_plot}
\alias{feat_plot}
\title{Feature plot}
\usage{
feat_plot(object, Features, raster = NULL, cells = NULL, order = NULL)
}
\arguments{
\item{object}{The Seurat object}

\item{Features}{Vector containing the names of the selected features}

\item{raster}{Convert points to raster format (Default=NULL)}

\item{cells}{Select cells to be plotted (Default=NULL)}

\item{order}{Boolean determining whether to plot cells in order of expression (Default=NULL)}
}
\value{
List containing the feature plot of each selected feature
}
\description{
This function generates a list containing the feature plot of each selected feature.
Feature plot: Coloring single cells on a dimensional reduction plot according to a 'feature'
}
