% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_transition.R
\name{find_transition}
\alias{find_transition}
\title{Find transition points between two time points}
\usage{
find_transition(data, cluster, feat_num)
}
\arguments{
\item{data}{You data}

\item{cluster}{Name of the cluster column}

\item{feat_num}{Number of features}
}
\value{
A list that contains a correlation matrix and the two clusters with highest correlation, 
considered as transition points between the two time points.
}
\description{
This function finds out the transition points between two time points based on their correlation
}
