% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\name{Forecast}
\alias{Forecast}
\title{Forecasting future time steps}
\usage{
Forecast(data, batch_size = 25, train_length = 150, tsteps = 1, epochs = 500)
}
\arguments{
\item{data}{Your time step dataframe}

\item{batch_size}{The number of training examples in one forward/backward pass of a 
RNN before a weight update. It must be evenly divisible into both the training and testing lengths}

\item{train_length}{Number of training time steps}

\item{tsteps}{The number of lags included in the training/testing set}

\item{epochs}{The total number of forward/backward pass iterations}
}
\value{
A dataframe with both input and forecasted time steps
}
\description{
This function forecasts the future time steps of a selected feature
}
