% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_tk_data.R
\name{Gen_tk_data}
\alias{Gen_tk_data}
\title{Generate time step data}
\usage{
Gen_tk_data(data, cluster, Feature)
}
\arguments{
\item{data}{Your data}

\item{cluster}{The name of the cluster column}

\item{Feature}{Your selected feature}
}
\value{
A time step dataframe, where "index" are the time steps
}
\description{
This function takes your longitudinal dataframe and generates a time step dataframe for a selected feature
}
