% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/important_feat.R
\name{important_feat}
\alias{important_feat}
\title{Important features}
\usage{
important_feat(model, n = 30, cell_num, Decrease = TRUE, main = "")
}
\arguments{
\item{model}{The random forest model}

\item{n}{The number of features that will be selected (Default=30)}

\item{cell_num}{The number of cells}

\item{Decrease}{In which order should the features be selected? Increasing or decreasing?
(Default=TRUE)}

\item{main}{The title of the feature importance plot}
}
\value{
List containing the selected features and a feature importance plot
}
\description{
This function selects n features by their importance calculated by the random forest model
}
