% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter_corr.R
\name{iter_corr}
\alias{iter_corr}
\title{Iteration of random-subsets correlation matrix}
\usage{
iter_corr(data1, data2, n = 100, decreasing = TRUE)
}
\arguments{
\item{data1}{Time-point data.frame 1}

\item{data2}{Time-point data.frame 2}

\item{n}{Number of iterations (Default=100)}

\item{decreasing}{Show correlations decreasing? (Default=TRUE)}
}
\value{
A list which contains n data.frames of the correlation of each feature in different time-points
}
\description{
This function creates n random-subset correlation matrices between two data.frames
}
