% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_quality.R
\name{model_quality}
\alias{model_quality}
\title{Check quality}
\usage{
model_quality(model, test_set, real_output)
}
\arguments{
\item{model}{The random forest model}

\item{test_set}{Data which variable/factor/output is predicted by the model}

\item{real_output}{The real variable/factor/output of the test data}
}
\value{
List containing the predicted data, the confusion matrix, the model quality table and a heatmap representing
the correctness of the prediction.
}
\description{
This function checks the quality of the model
}
