% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dyn.R
\name{plot_dyn}
\alias{plot_dyn}
\title{Feature dynamics per individuals}
\usage{
plot_dyn(data)
}
\arguments{
\item{data}{Data}

}
\value{
A plot representing the dynamic change of feature and
factor expressions.
}
\description{
This function generates a plot representing the dynamics of the expression of one feature in different patients
}
