% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pred.R
\name{plot_pred}
\alias{plot_pred}
\title{Plot prediction}
\usage{
plot_pred(data, size = 1, base_size = 14)
}
\arguments{
\item{data}{The complete data (original + predicted)}

\item{size}{The width of the plotted line}

\item{base_size}{Bae font size}
}
\value{
The plot of the original time steps and the predicted time steps
}
\description{
This function plots the predicted data colouring it different to the original data
}
