% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tk.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Plot time step dataframe}
\usage{
plot_time_series(tk_df, Feature, dyn_col = "red", title = "")
}
\arguments{
\item{Feature}{Your seleted feature}

\item{dyn_col}{The color of the plot}

\item{title}{Title of the plot}

\item{data}{Your time step dataframe}
}
\value{
A plot showing the dynamics of the feature
}
\description{
Thus function generates a plot showing the dynamics of the selected feature
}
