% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDecTree.R
\name{plotDecTree}
\alias{plotDecTree}
\title{Plot the decision tree of the model}
\usage{
plotDecTree(
  model_type = c("rf", "tree"),
  model,
  tree_num = 2,
  num_tree = 200,
  main = "Decision tree",
  data,
  formula,
  type = 0
)
}
\arguments{
\item{model_type}{Select which model do you want to use (rf=randomForest, tree=tree model)}

\item{model}{The model. If no model provided, a selected (rf or tree) will be generated}

\item{tree_num}{The index of the tree which you want to extract from the model. Only when 'model_type="rf"'}

\item{num_tree}{The number of trees you want to generate.
Only when 'model_type="rf"' and no model is provided (Default=200)}

\item{main}{The title of the plot}

\item{data}{The data that will be used to generate the model. Only when no model is provided}

\item{formula}{The formula to generate the model. Only when no model is provided}

\item{type}{The type of plot that you want to generate from the tree model. Only when 'model_type="tree"'}

\item{generate}{Should a}
}
\value{
A decision tree
}
\description{
This function plots a decision tree, it can be extracted from the random forest model, or
from generating a decision tree model.
}
