% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest.R
\name{RandomForest}
\alias{RandomForest}
\title{Generate random forest model}
\usage{
RandomForest(
  formula,
  data,
  split = TRUE,
  ntree = 200,
  strata = NULL,
  importance = TRUE,
  proximity = TRUE
)
}
\arguments{
\item{formula}{Formula describing the model to be fitted}

\item{data}{The dataframe containing the variables in the model}

\item{split}{Train Test Split ? (Default=TRUE)}

\item{ntree}{Number of trees to grow}

\item{strata}{A variable that is used for stratified sampling (Default = NULL)}

\item{importance}{Should predictors' importance be assessed?(Default=TRUE,
recommended for further analyses)}

\item{proximity}{Should proimity measure among the rows be calculated? (Default = NULL)}
}
\value{
The random forest model with its input data
}
\description{
This function generates a random forest model
}
