% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_tree_func.R
\name{rf_tree_func}
\alias{rf_tree_func}
\title{Plot extracted decision tree from forest}
\usage{
rf_tree_func(model, tree_num, main)
}
\arguments{
\item{model}{The random forest model}

\item{tree_num}{The index of the tree which you want to extract from the model}

\item{main}{The title of the plot}
}
\value{
A decision tree
}
\description{
This function plots a decision tree extracted from a random forest model.
}
