% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_sampling1.R
\name{sub_sampling1}
\alias{sub_sampling1}
\title{Get frequency}
\usage{
sub_sampling1(input, perc = 0.1)
}
\arguments{
\item{input}{Input data}

\item{perc}{The fraction of cells selected in each cell type}
}
\value{
A data.frame that contains the number of cells from each cell type.
}
\description{
This function calculates the number of cells which will be randomly selected from each cell type
in the next step
}
