% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_sampling2.R
\name{sub_sampling2}
\alias{sub_sampling2}
\title{Sub-sampling}
\usage{
sub_sampling2(Freq, data)
}
\arguments{
\item{Freq}{The number of samples that are selected from each group}

\item{data}{The data which will be sub_sampled}
}
\value{
A sub-sampled input data.frame
}
\description{
This function select randomly n cells from each cell_type according to the input frequencies
}
