% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Within_variance.R
\name{Within_variance}
\alias{Within_variance}
\title{Significance of the within time-point variance}
\usage{
Within_variance(data1, data2, n = 100)
}
\arguments{
\item{data1}{Time-point data.frame 1}

\item{data2}{Time-point data.frame 2}

\item{n}{Number of iterations (Default=100)}
}
\value{
A data.frame containing the p-values between the iterations
}
\description{
This function calculates the significance of the variance within time-points
}
