% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactFeat_Dyn.R
\name{FactFeat_Dyn}
\alias{FactFeat_Dyn}
\title{Factor-Feature dynamics}
\usage{
FactFeat_Dyn(data, Features)
}
\arguments{
\item{data}{Entire data.frame}

\item{Features}{The selected features}
}
\value{
A plot representing the dynamic change of feature and
factor expressions among time-points.
}
\description{
This function generates a plot representing the dynamic change of feature and
factor expressions among time-points.
}
