% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactFeat_ExprRegr.R
\name{FactFeat_ExprRegr}
\alias{FactFeat_ExprRegr}
\title{Factor-Feature expression regression}
\usage{
FactFeat_ExprRegr(
  Feat_data,
  Fact_data,
  Cell_Type,
  Time_Step = NULL,
  Feature = "",
  main = ""
)
}
\arguments{
\item{Feat_data}{Data.frame of feature expressions}

\item{Fact_data}{Data.frame of factor expression}

\item{Cell_Type}{Cell-type of each input cell}

\item{Time_Step}{Split the plot by time-points? If 'yes', input the time-point of each input cell.
Otherwise, ignore this parameter.}

\item{Feature}{Which feature do you want to check?}

\item{main}{The title of the plot}
}
\value{
A plot comparing the expression of the feature and the factor.
}
\description{
This function generates a plot comparing a feature with a factor
}
